Undo rename of vibration
