+++ src/api/VibrationIOS.js
